Revised version of dataCleaning script
added NULL values for playTime missing in the last level students
played.
added NULL values for cases in which badgeDecision did not apply.

changes to consider in the future:
maybe we will need to go one level further down (at least in the CSV,
not necesarilly in the tool) after LEVEL, store information for each
playCount. As it is right now we are losing information and what gets
stored is the value for the last time the student played a certain
level....